Support running in development mode
As well as building a static bundle with `npm run build`,
create-react-app supports running a development server (with unminified
JS) via `npm start`. This commit adds a new Dockerfile.dev to allow
docker-compose to set the port that the development server will use, and
removes a problematic environment variable that breaks everything if
you're not running the development server at localhost:8000.