Guard against seemingly impossible situation
Sometimes Axios appears to execute an HTTP request even though the
adapter which executes the request has been replaced, which results in
the caching layer getting confused and spitting out errors into the
console (and on top of the application in development mode, which makes
debugging quite irritating). Fixing this -- or at least finding out
what's really happening -- would be nice, but it doesn't seem to be
causing any problems, so we'll just log that something strange is going
on for now.