Remove an extraneous call to requestRotations()
requestRotations(1) should be called once for each call to
receiveRotations() (or, equivalently, requestRotations(n) should be
called once for every n calls to receiveRotation()), but in this
function it was called even though there was no matching call to
receiveRotations(). This led to the number of "fetching" rotations
increasing without bound as the app was interacted with -- this doesn't
currently cause any issues as that number isn't used anywhere, but could
in the future.