Cope with missing data in user filter
Frequently parts of the data for each user will be mising (for example,
the personal email), in which case `str` will be `null`, which would
previously cause the search to throw an error and break the entire page
when you entered anything at all.

Strictly speaking, this should be a ternary checking for `str === null`,
but it doesn't matter that it also fires on `str === ""`, since a
non-empty `textFilter` will never be contained by the empty string.