Make the App component use Redux like everything else
Previously, App was using the Redux store directly, unlike the rest of
the application (which used it via React Redux), even though it was
connect()ed and had mapState and mapDispatch functions. Using the store
via this.props instead of directly improves consistency with other
components, and apparently allows React Redux to apply some performance
optimisations.