Fix issue where the app wouldn't start
Since App.componentDidMount no longer contains a call to authenticate(),
there was nothing to trigger a state change and cause componentDidUpdate
to be called, so the app would just sit there and do nothing. Fetching
some data in componentDidMount fixes this (probably it's not necessary
to keep the calls in componentDidUpdate around any more too).