Always fetch the entire latest series
Currently, App only fetches the latest rotation; whilst this is
technically "correct", in that App only cares about getting the ID of
the latest rotation, it's unhelpful for the rest of the application,
most of which cares about the whole of the latest series. In particular,
the homepage (provided mostly by MainPage) displays all rotations from
the latest series, and since it only fetches the latest series once it
has been mounted, it currently renders once while only displaying the
latest rotation, and then updates just after with all rotations from the
latest series.

Ideally, App would mount MainPage even when there is no latest rotation,
so that MainPage can fetch the latest series at the same time as App is
fetching the latest rotation, but that would require more significant
refactoring of the application to cope with the case that sometimes
there is no latest rotation.