Hide "Create Rotation" button for even longer
The current behaviour of showing the button as soon as the latest
rotation becomes uploadable has existed since the start of the project
(15d825a, "Initial Commit"), but it's important that a new rotation is
not created until the previous rotation is no longer useful, so we gate
it on project assignments having been finalised as well.

Arguably this should actually be gated on the latest rotation's
marking_complete date having passed, not because the backend's scheduled
job for that deadline does anything, but because that's a better
estimate of when the rotation stops being useful to look at. It would
probably be bad if a new rotation was created while students were still
uploading their reports or reading feedback from markers, for example.