Allow any date to be picked for rotation deadlines in developer mode
It's bad if users can change dates from being in the future to in the
past, since this can cause scheduler events to be missed or (presumably)
happen out of order. However, it's useful for development purposes to be
able to create and manipulate rotations far in the past (more than
misfire_grace_time, which is currently 31 days), so that the scheduler
doesn't run events at all.