Fix misbehaviour when adding users
If the empty "new user" row had the user type edited, and then all roles
removed again, the row wouldn't disappear. This is because in
JavaScript, `[] !== []`, and the equality-testing library(!) that had
been chosen was specifically for testing shallow equality, which is not
the same (ha) as actually useful "are these things equivalent" equality.