Actually genuinely fix project list sorting, for real this time
This sorting code has been through six iterations:

- fd450a4 ("Only show vote buttons when needed") -- sorts ascending by
  last name
- 64c79aa ("Sort project list better") -- sorts ascending by whether
  this user is the assigned student, then by whether this user can mark
  this project, then by last name, then by project title, which has the
  effect that the projects you are assigned to or can mark are pinned to
  the bottom of the list (since true is greater than false)
- 5da52d7 ("Show markable projects properly") -- corrects the error in
  the previous commit, so assigned/markable projects are now pinned to
  the top of the list
- 9cfe231 ("Sort project lists better") -- sorts ascending by
  unassignedness, then by unmarkability, then descending by rotation,
  then ascending by last name, then by title
- f269a86 ("Sort project lists better") -- sorts like the previous
  commit, but now ascending by rotation (not descending)
- e22a4e2 ("Finally fix sorting") -- sorts like the previous commit, but
  descending (not ascending), and using a much more complicated method
  that considers each criterion in turn rather than all at once

This commit reverses the change from ascending to descending that the
last commit made, but reverses also the direction in which projects are
sorted by rotation. In summary, projects are now sorted, in order, thus:

- descending by whether the user is assigned to this project, therefore
  assigned projects are pinned to the top of the list
- descending by whether the user is able to mark this project, therefore
  markable projects are pinned to the top of the list
- descending by rotation ID (which *currently* correlates with the
  progression of time), therefore projects from newer rotations are
  shown above projects from older rotations
- ascending by last name, therefore names closer to the start of the
  alphabet come before names closer to the end of the alphabet
- ascending by project title, with similar effect

This seems like roughly the sum total of the previous six commits'
intentions, but without the implementation bugs.

This sorting code also contained a bug whereby no project would be
considered assigned to this user, which this commit fixes.

(Why does this component have to worry about showing projects from
multiple rotations? Because it's used everywhere a list of projects is
shown, including in MainPage, where a list of all projects the user is
assigned to is shown to each student. It would therefore be good if the
project from the most recent rotation could be shown at the top of that
list.)