Remove use of Object.fromEntries
This is new in ES2019, and is not polyfilled or transpiled by any of the
existing tooling (and browser support is not good enough to use it
without a polyfill). It's easier to rewrite it in terms of simpler ES6
primitives, which *are* transpiled, than to find and install a polyfill.