Unify the two CoGS marker editing pages
The only difference between these two pages is the buttons shown ("Save
and Return" and "Finish" before projects are finalised, vs. "Save
Changes" afterwards), so it's probably easier to just make them into the
same page.

Instead of being located at /rotations/choices/cogs or
/rotations/:part/cogs, this page is now located at
/rotations/:series/:part/cogs, which avoids some of the weirdness that
was previously going on to infer what the latest series was.

The Finalise Projects button might be shown on another page in the
future, so having it as its own component is handy.