Unify RotationChoiceViewer and RotationChoiceChooser
These two pages are almost identical, and in fact they render the same
component, just with slightly different props. Combining them adds
little complexity to the code; the only feature lost is the different
title for viewing and finalising project assignments.

Also, the "View Student Choices" link is now displayed whenever the last
rotation is viewable by students, not only when students can make
project choices (unless the rotation is waiting to be finalised, in
which case the button is replaced with the "Finalise Student Choices"
button).

Given the new URL structure, it would be reasonable to assume that this
page works correctly for previous rotations too -- however, it does not.
It displays one row for each *current* student, which is not a problem
for past rotations in the current series, but students from past series
will most likely no longer have the student role, so they will not be
displayed (instead, students from the current series will be shown).
Fixing this would be nice (but the only link exposed in the UI is the
link to the latest rotation, so it's not necessary right now).