Hide "View Student Choices" page when choices have been erased
When project assignments are finalised, each student's choices are
blanked (to prepare for the next rotation), which means that this page
becomes useless. This does mean that currently the only way for the
Graduate Office to view which student is assigned to which project is
via the exported spreadsheets, which is unfortunate.

This reverts part of commit 4f49d74 ("Unify RotationChoiceViewer and
RotationChoiceChooser").