Fix incorrect usages of this.setState
Calls to setState can be batched and executed later, so this.props and
this.state should not be used in arguments to setState. Also, setState
implicitly does a shallow merge of the new state with the previous
state, so a large number of the calls to immutability-helper's update()
were unnecessary.