Work around broken checkbox behaviour in MultiselectDropDown
The problem: if the user clicked on a checkbox in the
MultiselectDropDown (rather than clicking on the text adjacent to the
checkbox), the state of the checkbox would appear visually to stay the
same, but everything else -- its props, the parent element's state, etc.
-- correctly believed that the checkbox has been toggled.

To fix this, we force an update of the entire component every time a
checkbox is clicked, which rerenders the checkbox and results in the
visual appearance matching the React/DOM state again. Both the call to
setTimeout and the two arrow functions are required; without the arrow
functions, various React code doesn't get the correct `this`, and
without the setTimeout, it Just Doesn't Work.

It's unclear exactly why this works. There are several issues in the
React repo and elsewhere describing similar behaviour:

facebook/react#2766
facebook/react#3005
facebook/react#3446
react-bootstrap/react-bootstrap#2485

Unfortunately, none of the suggestions made appear to be directly
applicable to this situation.

This workaround is bad for two reasons:
- it's ugly and nonobvious
- it forces a rerender of far too many components

Nevertheless, it fixes the problem.

Closes #1.