Update list of programmes
This shouldn't cause issues elsewhere in the application; the backend
treats the list of programmes as a list of opaque strings that do not
contain the character `|`, and the rest of the frontend typically
constructs a mapping from programme to boolean based on this constant,
not from the programmes persisted in the database, so edited or
resubmitted projects should automatically have the old programmes
removed.

Fixes #2.