Correct button layout in GroupForm
On sm and md, the save button and the remind button would stack up on
top of each other in an ugly way, and this would even happen on lg if
the text in either button was even slightly longer. This is because
previously, the buttons were part of the same column as the deadline
date fields.

This commit moves the buttons into their own row, since there is no
reason to constrain them to the deadlines column, and also adjusts the
widths of the two existing columns to better reflect their contents
(previously, the columns overlapped, which would have caused issues if
the text in either column became longer).

Now, the buttons are aligned left and right at all sizes, with 1 column
of padding to the sides at sm+ (to match the other content).