Fix user priorities of 0 not being shown
Although it would be nice not to display a priority for users who are
not students, it's more important not to hide priorities for users who
got their first choice, as this is very confusing.

userDefault is compared against the raw field values, before the
priority is parsed to an integer, so it must contain a string also.