Show more detail in finalise choices screen
The supervisor and project type (computational/wetlab) are now shown
after each project title.

This also adds support for a class ".btn-long" which can be used to
modify the styling of a button that is expected to contain long text,
which makes the button's content wrap over multiple lines.

See #14.