Prevent a project's supervisor being assigned to be its CoGS marker
Just disabling the menu item has the potential to cause confusion,
especially since the disabled style is not particularly visible, but
removing the user from the list entirely (and thereby changing the
contents of the list for every project) would /probably/ be worse.
Hopefully the Graduate Office won't try to select the same user to be a
project's supervisor and CoGS marker anyway.

A title attribute probably wouldn't help, since they're not particularly
discoverable even to experienced users.

Fixes #15.