Make all actions return promises
Next step: swapping out all of the ad-hoc callbacks and replacing them
with .then() and/or .catch() as appropriate.

Promise.allSettled currently has very poor browser support (Chrome 76 is
the only desktop browser to support it, though Firefox has implemented
it and may ship soon[1], and only Chrome and Opera support it on
mobile), so we use a polyfill to ensure compatibility with other
widely-used browsers.

[1]: https://bugzilla.mozilla.org/show_bug.cgi?id=1549176