Link to GitHub wikis in documentation panel
Put the contents in a row (with the user guides and the technical
documentation each in a col-sm-4) to take up a little less vertical
space on desktop. (Possibly worth putting them in a flex container with
flex-grow: 1 and flex-shrink: 0 instead, but the Bootstrap way is easier
and less likely to look broken on mobile.)

Fixes #20.