Fetch only students in ProjectEditor
Commit c6546c7 ("Fetch users in project editor") overzealously fetched
all users when this component was mounted; whilst this resulted in the
correct behaviour, each user fetched results in a network request, so
it's ideal to keep the number of users requested as low as possible.
This is achieved by only asking the backend for users who can actually
join a project.