Fix broken comparison functions in sorts
It's unclear what the original intent behind some of these comparison
functions was; the ECMAScript spec is fairly unambiguous that
Array.prototype.sort() comparison functions must be consistent (for any
pair (a, b), the function must always return the same class of value
(positive, negative, or zero)), but comparing two values with the
greater-than or less-than operators does not a comparison function make.

Users in the CoGS marker list dropdown are now sorted ascending by
surname; the student choice viewer/finalisation screen is sorted
descending by student priority; and the main page (seriously, what was
the deal with `.sort(rotation => rotation.data.part`!?) is now sorted
*descending* by rotation part, so the most recent rotation appears at
the top of the page.