Support changing the supervisor of a project
Only users with the modify_permissions permission (that is, members of
the Graduate Office) should be allowed to change the supervisor
associated with a project. Users who are not part of the Graduate Office
can only create projects where they are the supervisor.

See wsi-cogs/backend@dfb3a91.

(I really wish this version of Create React App supported the nullish
coalescing/optional chaining operators...)