Use ProjectFeedbackForm for both reading and writing feedback
Previously, the ProjectFeedbackForm was only used for reading feedback.
The ProjectMark page shared the vast majority of its code, but had some
slight differences -- for example, in ProjectMark, grades were indexed
by numerical ID, and in ProjectFeedbackForm, the state contained HTML
strings (it now contains opaque react-rte "values", like how ProjectMark
used to work).

Now, ProjectFeedbackForm is used for both reading and writing feedback.
ProjectMark now renders a ProjectFeedbackForm instead of using
copy-pasted code.

To simplify the code, the backend API for submitting feedback now
accepts a string representing the grade instead of a numerical ID, so
this change also modifies the frontend to send the name of the grade
rather than its index.

This change must be deployed at the same time as the accompanying
backend change ("Use strings rather than numbers for grades in the
API"), for obvious reasons.