Display which users a choice conflicts with on the finalisation screen
Previously, if a conflicting choice was made, the two offending rows
would turn BRIGHT RED, but no other feedback was visible -- since the
other row would frequently be scrolled off the screen, this is not
exactly the most helpful way the application could tell the user that
something is wrong ("you did something wrong, but I'm not going to tell
you what it is").

Now, the name of the user(s) with which a choice conflicts is/are
displayed on the right-hand side of the table. That space is reserved
even if there is currently no conflict, since having the table jump
around underneath you if you choose a conflicting project is even worse
than the existing behaviour!

See #14.