Fix Dockerfiles by pinning to Node 6
Node 6 is atrociously outdated (it's so old that the Nodesource docs
don't even mention how to install it any more!), but it still works for
our purposes. The nginx Docker image changed from Debian Stretch to
Debian Buster a little over a month ago[1], which means that the version
of Node provided by the built-in repositories changed from 4.8 to 10, so
we were accidentally getting Node 10 instead of Node 6, since apt
was installing the highest version available.

Node 10 alone would probably have been fine; the problem was that we
were getting it from the Debian repositories, which package Node and npm
separately, whereas the Nodesource repositories bundle them together.
Therefore, the subsequent `npm install` was failing, because npm wasn't
installed!

[1]: nginxinc/docker-nginx@7956a3f