Upgrade npm in Dockerfiles
Node 6 comes with npm 3 by default, which is (unsurprisingly) just as
old and outdated as Node 6. It's so old that it doesn't understand
package-lock.json, so the dependencies there were being totally ignored
when deploying the application. (The only reason it might have appeared
to work is if you had a node_modules folder that you'd created with a
newer version of npm, which was being copied into the Docker container.)

Thankfully, current versions of npm still support Node 6. When that
changes it seems likely that we'll have to upgrade the version of Node
we use.