Fix possible bug with Finalise Choices screen
The RotationChoiceEditor's `projects` prop contains all projects in the
state; if the state contained projects from an old rotation (e.g. the
user had previously been browsing the list of projects from a previous
rotation), then it's possible that a student would be (visually)
assigned to a project which they had been assigned to in a previous
rotation -- if the user then pressed save again, this could potentially
have been persisted into the database!

It's possible that subtle features of the ES2017 spec would save us
here (by guaranteeing that integer keys are iterated over in order), but
that depends on who you ask, and it seems likely that some browsers
misimplement the spec anyway.