Don't allow students to be edited once they are "in use"
This is a nasty, hacky heuristic to avoid the problem of "Graduate
Office adds new user by replacing old user who's left Sanger".

This doesn't work if someone who isn't a supervisor leaves; the real
solution is probably to put a bin icon at the right-hand side of each
row, which just archives that user.

Fixes #4.