Add information on emails & priority to user edit page
From user testing, it's clear that users are confused about why there
are two fields for email addresses, and what the point of the "priority"
field is. Hopefully these new tooltips/popovers will clarify what each
of them means.

The new InfoButton component listens for click, focus, and hover events;
ideally it would only listen for clicks, because currently (due to
having them also be rootClose) they open and then immediately close if
you try to click on the icon (since clicking on it implies hovering over
it first). However, a click handler alone appears to be
keyboard-inaccessible, which is not great. Hopefully a user who clicks
the icon immediately will realise that they can click it again, or move
the cursor away and back, to show the popup.