Don't display a bogus rotation status for the initial rotation
When the backend is brought up for the first time, an initial dummy
rotation is created; this rotation is invisible to students (to minimise
confusion), but is also read-only (to stop supervisors creating projects
in it). The Graduate Office needs to create a new rotation rather than
modifying this one.