Remove the "upload final project" header link
Since this uses the student's latest assigned project to determine
whether or not it should be displayed, and students can be assigned to
new projects before they have uploaded reports for their previous
projects, this button would disappear as soon as the student's next
project was assigned, which would be incredibly confusing (since it also
disappears when you're no longer allowed to upload).

This was a useful link to have, and it would be nice to resurrect it at
some point, but there's no immediately obvious way to do so right now.