Update the Project Status (which includes the names of uploaded files) within the application state when new files are uploaded.
The names of files uploaded are stored as part of application state. Earlier, when new files were uploaded to overwrite old ones, the names only showed old file names as the state wasn't updated. Technically, there are two ways to update the state after a succesful upload: we could use the filenames which were stored in the local component state in a separate attribute (uploads[]), or we could grab the names from the response of the project.upload API. I opted for the latter to keep them coupled with the remote backend.

Resolves: Issue #31